--- a/src/test/java/com/hedera/mirror/dataset/AccountBalancesFileLoaderIT.java
+++ b/src/test/java/com/hedera/mirror/dataset/AccountBalancesFileLoaderIT.java
@@ -9,9 +9,9 @@ package com.hedera.mirror.dataset;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -37,10 +37,10 @@ public class AccountBalancesFileLoaderIT {
         final var fileName = "2019-08-20T21_30_00.147998006Z_Balances.csv";
         final var path = getClass().getResource(Paths.get("/account_balances", fileName).toString());
         final var cut = new AccountBalancesFileLoader(Paths.get(path.toURI()));
-        cut.loadAccountBalances();
+        boolean success = cut.loadAccountBalances();
         assertAll(
                 () -> assertEquals(2, cut.getValidRowCount())
-                ,() -> assertFalse(cut.isInsertErrors())
+                ,() -> assertTrue(success)
         );
         // TODO assert the rows actually added to the database.
     }