<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Name your PR as `<Feature Area>: Describe your change`
3. Rebase your PR if it gets out of sync with master
-->

**Detailed description**:
- Removed accountBalancesUseTransaction. Logs below show no perf difference,
  so keeping it simple by removing db transaction option.
- There was partial error handling in processRecordStreamWithoutDatabaseTransaction.
  Errors in last batch  would have propagated up and skipped remaining logic
  in loadAccountBalances(). Fixed.
- insertError (only needed in test) didn't need to be state variable.
  Just passing it as return value was enough. One less state variable.
- Minor logging improvement

**Which issue(s) this PR fixes**:
Fixes #297
Fixes #117

**Special notes for your reviewer**:
Comparing performance (running locally on laptop):

_Without Database txn_
2019-10-04 16:23:10,697 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T12_00_00.040026Z_Balances.csv with 53667 records in 25.17 s
2019-10-04 16:23:35,648 INFO  [task-1] …balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T12_45_00.036906001Z_Balances.csv with 53669 records in 24.95 s
2019-10-04 16:24:09,066 INFO  [task-1]… balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T08_00_00.033413001Z_Balances.csv with 53634 records in 33.44 s
2019-10-04 16:24:36,983 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T18_15_00.015393Z_Balances.csv with 53883 records in 27.92 s
2019-10-04 16:25:02,511 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T08_45_00.017710Z_Balances.csv with 53644 records in 25.52 s
2019-10-04 16:25:05,825 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T02_30_00.029959Z_Balances.csv with 53511 records in 3.312 s
2019-10-04 16:25:29,676 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T15_15_00.023209001Z_Balances.csv with 53693 records in 23.85 s

_With database txn_
2019-10-04 16:28:50,700 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T23_15_00.042208001Z_Balances.csv with 54053 records in 21.31 s
2019-10-04 16:29:11,515 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T21_45_00.032315Z_Balances.csv with 54053 records in 20.81 s
2019-10-04 16:29:14,652 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T11_45_00.012213Z_Balances.csv with 53667 records in 3.135 s
2019-10-04 16:29:38,843 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-04T22_00_00.055866Z_Balances.csv with 54053 records in 24.19 s
2019-10-04 16:30:08,299 INFO  [task-1] … balances file ./MirrorNodeData/accountBalances/valid/2019-10-03T23_30_00.015795001Z_Balances.csv with 53511 records in 29.45 s